updated tracker to return the dto
